requirements.txt
absl-py==2.3.1
accelerate==1.11.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.1
aiosignal==1.4.0
altair 
annotated-types==0.7.0
anyio==4.6.2.post1
astunparse==1.6.3
async-timeout==5.0.1
attrs 
beautifulsoup4==4.13.3
blinker 
Bottleneck 
brotlicffi 
cachetools 
certifi 
cffi==1.17.1
chardet==5.2.0
charset-normalizer 
click 
colorama 
contourpy==1.3.2
cryptography==43.0.3
cycler==0.12.1
datasets==4.2.0
dill==0.4.0
distro==1.9.0
editdistpy==0.1.6
emoji==2.15.0
exceptiongroup==1.3.0
filelock==3.20.0
flatbuffers==25.9.23
fonttools==4.60.1
frozenlist==1.8.0
fsspec==2025.9.0
gast==0.6.0
gensim==4.4.0
gitdb 
GitPython 
google-pasta==0.2.0
grpcio==1.76.0
h11==0.14.0
h5py==3.15.1
httpcore==1.0.6
httpx==0.27.2
huggingface-hub==0.35.3
idna==3.11
Jinja2 
jiter==0.6.1
joblib==1.5.2
jsonschema 
jsonschema-specifications 
keras==3.11.3
kiwisolver==1.4.9
libclang==18.1.1
Markdown==3.9
markdown-it-py 
MarkupSafe 
matplotlib==3.10.7
mdurl 
mkl-service==2.4.0
mkl_fft 
mkl_random 
ml_dtypes==0.5.3
mpmath==1.3.0
multidict==6.7.0
multiprocess==0.70.16
namex==0.1.0
narwhals 
networkx==3.4.2
nltk==3.9.2
numexpr 
numpy 
openai==1.52.0
opt_einsum==3.4.0
optree==0.17.0
outcome==1.3.0.post0
packaging 
pandas 
pillow 
propcache==0.4.1
protobuf 
psutil==7.1.1
psycopg2==2.9.10
psycopg2-binary==2.9.10
pyarrow==21.0.0
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
pydeck 
Pygments 
pyparsing==3.2.5
PySocks 
python-crfsuite==0.9.11
python-dateutil 
python-dotenv==1.0.1
pytz 
pyvi==0.1.1
PyYAML 
referencing 
regex==2025.10.23
requests 
rich 
rpds-py 
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.15.3
seaborn==0.13.2
selenium==4.37.0
six 
sklearn-crfsuite==0.5.0
smart_open==7.4.1
smmap 
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.8
streamlit 
sympy==1.14.0
symspellpy==6.7.8
tabulate==0.9.0
tenacity 
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
termcolor==3.1.0
tf_keras==2.20.1
threadpoolctl==3.6.0
tokenizers==0.22.1
toml 
torch==2.9.0
tornado 
tqdm==4.66.5
transformers==4.57.1
trio==0.31.0
trio-websocket==0.12.2
typing_extensions 
tzdata 
underthesea==6.8.4
underthesea_core==1.0.4
urllib3 
vncorenlp==1.0.3
watchdog 
webdriver-manager==4.0.2
websocket-client==1.9.0
Werkzeug==3.1.3
win-inet-pton 
wrapt==2.0.0
wsproto==1.2.0
xxhash==3.6.0
yarl==1.22.0
